requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==4.2.26
django-allauth==65.13.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
jwt==1.4.0
psycopg2-binary==2.9.11
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
